Correct the removal of ACEs
We should only flush the permissions if the user asks us to with `--posix-only`. This changes the default behavior to "no", don't flush them. And, it changes the if statement to include the flush command -D. 

Also the terminology was clarified to ACEs, which matches how they are described in OneFS `man chmod`.