Add HTTP user to all Hive to proxy
In Ambari 2.4, the Hive service check when Kerberized verifies hadoop.proxyuser.HTTP.groups
by having the SPNEGO principal attempt to act on behalf of ambari-qa.

So, an HTTP user is needed on OneFS. And then a proxyuser needs to be configured. This
takes care of the first part. The implementation guide needs to be updated for proxyuser.